requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
torch>=2.1
omegaconf>=2.3
loguru>=0.7
mlflow>=2.8
streamlit>=1.27
plotly>=5.17
pytest>=7.4
ydata-profiling>=4.6
shap>=0.45
lime>=0.2
matplotlib>=3.7
pillow>=9.5
jinja2>=3.1
weasyprint>=62.0
statsmodels>=0.14
optuna>=3.4
ray>=2.9
fastapi>=0.105
uvicorn>=0.24
pyyaml>=6.0
networkx>=3.1
